- Forgot to commit these for MSVC build...
svn path=/trunk/; revision=20661

+++ reactos/drivers/net/tdi/cte/string.c
@@ -9,8 +9,8 @@
 BOOLEAN
 STDCALL
 CTEAllocateString (
-	DWORD	Unknown0,
-	DWORD	Unknown1
+	ULONG	Unknown0,
+	ULONG	Unknown1
 	)
 {
 	return FALSE;
@@ -23,8 +23,8 @@ CTEAllocateString (
 BOOLEAN
 STDCALL
 CTEInitString (
-	DWORD	Unknown0,
-	DWORD	Unknown1
+	ULONG	Unknown0,
+	ULONG	Unknown1
 	)
 {
 	return FALSE;
